+++ header.d.ts
@@ -1,0 +1,7 @@
+/* Side-effect module: optional responsive navigation for <m-header>. */
+declare global {
+  interface HTMLElementTagNameMap {
+    "m-header": HTMLElement;
+  }
+}
+export {};